requirements.txt
Django>=4.0
beautifulsoup4>=4.12
requests
djangorestframework
lxml
django-cors-headers
dj-database-url 
psycopg2-binary

# image search API
ddgs

# for HTTPS connection
django-extensions
# werkzeug

pyOpenSSL


